Combine the Turbo and Strada configs into a single Hotwire config
