--- core/src/main/kotlin/dev/hotwire/core/bridge/Strada.kt
@@ -1,10 +1,0 @@
-package dev.hotwire.core.bridge
-
-object Strada {
-    val config: StradaConfig = StradaConfig()
-
-    fun userAgentSubstring(componentFactories: List<BridgeComponentFactory<*,*>>): String {
-        val components = componentFactories.joinToString(" ") { it.name }
-        return "bridge-components: [$components]"
-    }
-}
