+++ core/src/main/kotlin/dev/hotwire/core/bridge/StradaJsonConverter.kt
@@ -1,5 +1,6 @@
 package dev.hotwire.core.bridge
 
+import dev.hotwire.core.config.Hotwire
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 
@@ -13,7 +14,7 @@ abstract class StradaJsonConverter {
                 "or use the provided KotlinXJsonConverter."
 
         inline fun <reified T> toObject(jsonData: String): T? {
-            val converter = requireNotNull(Strada.config.jsonConverter) { NO_CONVERTER }
+            val converter = requireNotNull(Hotwire.config.jsonConverter) { NO_CONVERTER }
 
             return when (converter) {
                 is KotlinXJsonConverter -> converter.toObject<T>(jsonData)
@@ -23,7 +24,7 @@ abstract class StradaJsonConverter {
         }
 
         inline fun <reified T> toJson(data: T): String {
-            val converter = requireNotNull(Strada.config.jsonConverter) { NO_CONVERTER }
+            val converter = requireNotNull(Hotwire.config.jsonConverter) { NO_CONVERTER }
 
             return when (converter) {
                 is KotlinXJsonConverter -> converter.toJson(data)
