--- core/src/main/kotlin/dev/hotwire/core/turbo/config/Turbo.kt
@@ -1,13 +1,0 @@
-package dev.hotwire.core.turbo.config
-
-object Turbo {
-    val config: TurboConfig = TurboConfig()
-
-    /**
-     * Provides a standard substring to be included in your WebView's user agent
-     * to identify itself as a Turbo Native app.
-     */
-    fun userAgentSubstring(): String {
-        return "Turbo Native Android"
-    }
-}
